Changed so that radians are used in gfx communication
git-svn-id: https://svn.torandi.com/rocket@59 71c61736-3055-43ec-afd9-7e59d617a297

+++ c_client/ship.h
@@ -8,7 +8,7 @@ struct ship_t {
 	char nick[32];
 	int x;
 	int y;
-	int a; //angle
+	double a; //angle
 	float s; //speed
 	bool attr[NUM_GFX_ATTR];
